API

Index

Networks.EdgePersistence

Networks.SimpleNetwork

Networks.addedge!

Networks.addvertex!

Networks.all_edges

Networks.all_vertices

Networks.edge_incidents

Networks.edge_neighbors

Networks.edge_type

Networks.hasedge

Networks.hasvertex

Networks.link!

Networks.nedges

Networks.nvertices

Networks.rmedge!

Networks.rmvertex!

Networks.unlink!

Networks.vertex_incidents

Networks.vertex_neighbors

Networks.vertex_type

Reference

Networks.EdgePersistence — Type

EdgePersistence

Trait for edge persitence in a Network. It defines the behavior of edges when a vertex is removed. The following traits are defined:

PersistEdges: edges are never removed implicitly.

RemoveEdges: edges are always removed implicitly.

PruneEdges (default): edges are removed if left stranded (i.e. no other vertex is linked with it).

Networks.SimpleNetwork — Type

SimpleNetwork <: AbstractNetwork

A network represented as an adjacency list. It is the translation of SimpleGraph from Graphs.jl to the Network interface.

Warning

This is mostly a example of compatibility with the Graphs.jl interface, but shouldn't be used as Graphs.SimpleGraph has proven to be problematic.

Networks.addedge! — Function

addedge!(graph, e)

Adds edge e to the graph.

Networks.addvertex! — Function

addvertex!(graph, v)

Adds vertex v to the graph.

Networks.all_edges — Function

all_edges(graph)

Returns the edges in the graph.

Networks.all_vertices — Function

all_vertices(graph)

Returns the vertices in the graph.

Networks.edge_incidents — Function

edge_incidents(graph, e)

Returns the vertices connected by edge e in graph.

Networks.edge_neighbors — Function

edge_neighbors(graph, e)

Returns the edges neighboring edge e in the graph.

Networks.edge_type — Function

edge_type(graph)

Returns the type of edges in the graph. Defaults to Any.

Networks.hasedge — Function

hasedge(graph, e)

Returns true if edge e exists in the graph.

Networks.hasvertex — Function

hasvertex(graph, v)

Returns true if vertex v exists in the graph.

Networks.link! — Function

link!(graph, v, e)

Links vertex v with edge e in the graph.

Networks.nedges — Function

nedges(graph)

Returns the number of edges in the graph.

Networks.nvertices — Function

nvertices(graph)

Returns the number of vertices in the graph.

Networks.rmedge! — Function

rmedge!(graph, e)

Removes edge e from the graph.

Networks.rmvertex! — Function

rmvertex!(graph, v)

Removes vertex v from the graph.

Networks.unlink! — Function

unlink!(graph, v, e)

Unlinks vertex v from edge e in the graph.

Networks.vertex_incidents — Function

vertex_incidents(graph, v)

Returns the edges connected to vertex v in graph.

Networks.vertex_neighbors — Function

vertex_neighbors(graph, v)

Returns the vertices neighboring vertex v in the graph.

Networks.vertex_type — Function

vertex_type(graph)

Returns the type of vertices in the graph. Defaults to Any.